fix: remove token_synthetic module
