+++ Cargo.toml
@@ -14,7 +14,6 @@ paste = { version = "1.0.15" }
 utils = { path = "modules/utils" }
 access_control = { path = "modules/access_control" }
 token_lp = { path = "modules/token_lp" }
-token_synthetic = { path = "modules/token_synthetic" }
 incentives = { path = "modules/incentives" }
 reentrancy_guard = { path = "modules/reentrancy_guard" }
 upgrade = { path = "modules/upgrade" }
