--- modules/token_synthetic/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "token_synthetic"
-version.workspace = true
-edition = "2021"
-publish = false
-rust-version.workspace = true
-
-
-[lib]
-crate-type = ["rlib"]
-doctest = false
-
-[dependencies]
-soroban-sdk = { workspace = true }
-utils = { workspace = true }
-
-[dev-dependencies]
-soroban-sdk = { workspace = true, features = ["testutils"] }
-utils = { workspace = true, features = ["testutils"] }
